# python/app/docs/swagger.yaml
swagger: '2.0'
info:
  title: API Gestionale Dipendenti
  description: Documentazione API per la gestione dei dipendenti
  version: "1.0"
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /fatture/:
    get:
      summary: Ottieni tutte le fatture
      tags:
        - Fatture

      parameters:
        - name: page
          in: query
          description: Numero della pagina da restituire
          default: 1
          required: false
          type: integer

        - name: per_page
          in: query
          description: Numero di fatture per pagina
          default: 10
          required: false
          type: integer

      responses:
        200:
          description: Lista di Dettagli
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: '#/definitions/Fattura'
              pagination:
                type: object
                properties:
                  per_page:
                    type: integer
                  total:
                    type: integer
                  pages:
                    type: integer
            example:
              current_page: 1
              items: [
                {
                  "cliente": 20,
                  "data_vendita": "2017-02-12",
                  "dettagli": [ ],
                  "id_fattura": "b100",
                  "totale": 1000,
                  "venditore": "VC01"
                },
                {
                  "cliente": 18,
                  "data_vendita": "2017-01-20",
                  "dettagli": [ ],
                  "id_fattura": "b101",
                  "totale": 2000,
                  "venditore": "VL02"
                }
              ]
              pagination:
                per_page: 10
                total: 150
                pages: 15
        404:
          description: Nessuna fattura trovata

    post:
      summary: Crea una nuova fattura
      tags:
        - Fatture
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Fattura'
      responses:
        201:
          description: Fattura creata
          schema:
            $ref: '#/definitions/Fattura'
        400:
          description: Dati non validi
        500:
          description: Errore durante la creazione
  /fatture/{id}:
    get:
      summary: Ottieni una fattura
      tags:
        - Fatture
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: Fattura trovata
          schema:
            $ref: '#/definitions/Fattura'
        404:
          description: Fattura non trovata

    put:
      summary: Aggiorna una fattura
      tags:
        - Fatture
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Fattura'
      responses:
        200:
          description: Fattura aggiornata
        400:
          description: Dati non validi
        404:
          description: Fattura non trovata
        500:
          description: Errore durante l'aggiornamento

    delete:
      summary: Elimina una fattura
      tags:
        - Fatture
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: Fattura eliminata
        404:
          description: Fattura non trovata
        500:
          description: Errore durante l'eliminazione
  /fatture/advanced/search:
    get:
      summary: Ricerca fatture per cliente, venditore o data
      description: Recupera una lista paginata di fatture filtrate per ID cliente, ID venditore o data_vendita
      tags:
        - Fatture

      parameters:
        - name: id_cliente
          in: query
          type: integer
          description: Filtra per ID Cliente
          required: false

        - name: id_venditore
          in: query
          type: string
          description: Filtra per ID Venditore
          required: false

        - name: data_vendita
          in: query
          type: string
          format: date
          description: "Filtra per data vendita (formato: YYYY-MM-DD)"
          required: false

        - name: page
          in: query
          type: integer
          description: "Pagina richiesta (default: 1)"
          required: false
          default: 1

        - name: per_page
          in: query
          type: integer
          maximum: 100
          description: "Numero massimo di elementi per pagina (max: 100)"
          required: false
          default: 10

      responses:
        200:
          description: Risultati della ricerca
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/definitions/Fattura'
              pagination:
                type: object
                properties:
                  page:
                    type: integer
                  per_page:
                    type: integer
                  total:
                    type: integer
                  pages:
                    type: integer

        400:
          description: Parametro non valido (es. data errata)

        404:
          description: Nessuna fattura trovata con i filtri dati
  /fatture/dettagli/categorie:
    get:
      summary: Restituisce le fatture che includono prodotti di più categorie
      description: |
        Endpoint che recupera tutte le fatture che hanno almeno due prodotti appartenenti a categorie diversi.
        Utile per analisi o report avanzati.
      tags:
        - Fatture

      responses:
        200:
          description: Lista di fatture con dettagli multi-categoria
          schema:
            type: array
            items:
              $ref: '#/definitions/FatturaDettaglioCategoria'
        404:
          description: Nessuna fattura trovata con prodotti di categorie diverse
        500:
          description: Errore interno del server
  /fatture/categorie:
    get:
      summary: Restituisce il totale delle fatture per categoria
      tags:
        - Fatture
      responses:
        200:
          description: Lista delle fatture totali per categoria
          schema:
            type: array
            items:
              $ref: '#/definitions/FatturaCategoriaCount'
        404:
          description: Nessuna fattura trovata
  /fatture/categoria/{categoria}:
    get:
      summary: Recupera le fatture che includono prodotti della categoria specificata
      description: |
        Endpoint che restituisce una lista paginata di fatture filtrate per categoria di prodotto.
        La categoria è passata come parametro nell'URL.
      tags:
        - Fatture

      parameters:
        - name: categoria
          in: path
          description: Categoria del prodotto (es. latte, carne, grano)
          required: true
          type: string

        - name: page
          in: query
          description: "Numero della pagina richiesta (default: 1)"
          default: 1
          required: false
          type: integer

        - name: per_page
          in: query
          description: "Elementi per pagina (default: 10, max: 100)"
          default: 10
          maximum: 100
          required: false
          type: integer

      responses:
        200:
          description: Lista paginata di fatture appartenenti alla categoria
          schema:
            $ref: '#/definitions/PaginatedFattureByCategoria'

        404:
          description: Nessuna fattura trovata per questa categoria

        500:
          description: Errore interno durante il recupero delle fatture
  /fatture/inserimento/fattura_dettagli:
    post:
      summary: Crea una nuova fattura con dettagli
      description: |
        Crea una fattura e i suoi dettagli in un'unica chiamata.
        Usa il campo `cf_cliente` per trovare l'id_cliente corretto.
      tags:
        - Fatture

      parameters:
        - name: body
          in: body
          description: Dati della fattura da creare
          required: true
          schema:
            $ref: "#/definitions/FatturaWithDettagliPost"

      responses:
        201:
          description: Fattura creata con successo
          schema:
            $ref: "#/definitions/FatturaWithDettagliOutput"
        400:
          description: Dati non validi
        404:
          description: Cliente non trovato per il codice fiscale
        500:
          description: Errore interno durante la creazione

#      security:
#        - bearerAuth: []
...
definitions:
...
  Fattura:
    type: object
    required:
      - id_cliente
      - data_vendita
      - totale
    properties:
      id_fattura:
        type: string
      id_cliente:
        type: integer
        description: ID del cliente a cui è associata la fattura
      id_venditore:
        type: string
        description: ID del venditore a cui è associata la fattura
      data_vendita:
        type: string
        format: date
        description: Data della fattura (YYYY-MM-DD)
      totale:
        type: number
        format: float
        minimum: 0
        description: Importo totale della fattura
  FatturaDettaglioCategoria:
    type: object
    properties:
      id_fattura:
        type: string
        example: "F001"
        description: "ID della fattura"

      data_vendita:
        type: string
        format: date-time
        example: "Tue, 06 Feb 2018 00:00:00 GMT"
        description: "Data dell'acquisto"

      numero_categorie:
        type: integer
        example: 3
        description: "Numero di categorie di prodotto nella fattura"

      totale:
        type: number
        example: 128.9
        description: "Totale importo della fattura"
    required:
      - id_fattura
      - data_vendita
      - numero_categorie
      - totale
  FatturaCategoriaCount:
    type: object
    properties:
      categoria:
        type: string
        example: "latte"
        description: "Nome della categoria di prodotto (es. latte, carne, grano)"
      count:
        type: integer
        example: 36
        description: "Numero di fatture che includono prodotti di questa categoria"
    required:
      - categoria
      - count
  FatturaWithDettagliOutput:
    type: object
    properties:
      id_fattura:
        type: string
      id_venditore:
        type: string
      id_cliente:
        type: integer
      data_vendita:
        type: string
        format: date
      totale:
        type: number
      dettagli:
        type: array
        items:
          type: object
          properties:
            prodotto:
              type: string
            quantita:
              type: integer
            costo:
              type: number
  PaginatedFattureByCategoria:
    type: object
    properties:
      current_page:
        type: integer
        example: 1
        description: "Numero della pagina attuale"

      items:
        type: array
        items:
          $ref: "#/definitions/FatturaWithDettagli"
        description: "Lista di fatture con dettaglio prodotti"

      per_page:
        type: integer
        example: 10
        description: "Numero di elementi per pagina"

      total:
        type: integer
        example: 63
        description: "Totale di fatture disponibili per questa categoria"

      pages:
        type: integer
        example: 7
        description: "Totale pagine disponibili"
  FatturaWithDettagli:
    type: object
    properties:
      id_fattura:
        type: string
        example: "R001"
        description: "ID della fattura"

      data_vendita:
        type: string
        format: date-time
        example: "2022-08-11T00:00:00Z"
        description: "Data dell'acquisto"

      totale:
        type: number
        example: 22.5
        description: "Importo totale della fattura"

      venditore:
        type: string
        example: "VC01"
        description: "ID del venditore"

      cliente:
        type: integer
        example: 12
        description: "ID del cliente"

      dettagli:
        type: array
        items:
          $ref: "#/definitions/DettaglioProdotto"
        description: "Prodotti nella fattura"
  FatturaWithDettagliPost:
    type: object
    properties:
      cf_cliente:
        type: string
        example: "RSSMRA90A01H501Z"
        description: "Codice fiscale del cliente"
      data_vendita:
        type: string
        format: date
        example: "2024-10-05"
      id_fattura:
        type: string
        example: "C001"
      venditore:
        type: string
        example: "VL02"
      dettagli:
        type: array
        items:
          type: object
          properties:
            prodotto:
              type: string
              example: "filetto"
            quantita:
              type: integer
              example: 3
            costo:
              type: number
              example: 10.6
          required:
            - prodotto
            - quantita
            - costo
  DettaglioProdotto:
    type: object
    properties:
      id_fattura:
        type: string
        example: "R001"
        description: "ID della fattura"

      prodotto:
        type: string
        example: "mozzarella"
        description: "Nome del prodotto"

    required:
      - id_fattura
      - prodotto
...
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Inserisci il token JWT nel formato: `Bearer {token}`"